+++ ItemSearch/BetterGuide/FavoritedRecipes.cs
@@ -119,7 +119,7 @@ public sealed class FavoritedRecipesPlayer : ModPlayer {
     }
     public bool ResetRecipeState(int recipe) => favoritedRecipes.Remove(recipe) | blacklistedRecipes.Remove(recipe);
 
-    public static void SortRecipes() {
+    public static void FilterAndSortRecipes() {
         List<int> favorited = [];
         List<int> others = [];
         List<int> blacklisted = [];
@@ -137,8 +137,10 @@ public sealed class FavoritedRecipesPlayer : ModPlayer {
         int index = 0;
         foreach (var recipe in favorited) Main.availableRecipe[index++] = recipe;
         foreach (var recipe in others) Main.availableRecipe[index++] = recipe;
-        foreach (var recipe in blacklisted) Main.availableRecipe[index++] = recipe;
+        if (!Configs.FavoritedRecipes.HideBlacklisted) foreach (var recipe in blacklisted) Main.availableRecipe[index++] = recipe;
         foreach (var recipe in unknown) Main.availableRecipe[index++] = recipe;
+        for (int i = index; i < Main.numAvailableRecipes; i++) Main.availableRecipe[i] = 0;
+        Main.numAvailableRecipes = index;
     }
 
     public readonly RangeSet favoritedRecipes = [];